+++ geopackage-map/src/main/java/mil/nga/geopackage/map/geom/PolygonOrientation.java
@@ -1,0 +1,21 @@
+package mil.nga.geopackage.map.geom;
+
+/**
+ * Orientation types of polygon exterior and hole rings
+ *
+ * @author osbornb
+ * @since 1.3.2
+ */
+public enum PolygonOrientation {
+
+    /**
+     * Counterclockwise orientation
+     */
+    COUNTERCLOCKWISE,
+
+    /**
+     * Clockwise orientation
+     */
+    CLOCKWISE
+
+}