+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * Google Map Shape visible setters
 * Google Map Shape Markers size and is empty methods
 * Close Polygons converting from Google Map Shapes to Geometries
+* Default and optional polygon counterclockwise and clockwise conversion orientations
 
 ## [1.3.1](https://github.com/ngageoint/geopackage-android-map/releases/tag/1.3.1) (02-02-2017)